start adding new CLI scripts for different functionalities
